requirements.txt
streamlit
pandas
plotly
python-dateutil
openpyxl
Pillow
bcrypt
streamlit-authenticator
